Module aws_sdk_cognitoidentityprovider::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AddCustomAttributes

Fluent builder constructing a request to AddCustomAttributes.

AdminAddUserToGroup

Fluent builder constructing a request to AdminAddUserToGroup.

AdminConfirmSignUp

Fluent builder constructing a request to AdminConfirmSignUp.

AdminCreateUser

Fluent builder constructing a request to AdminCreateUser.

AdminDeleteUser

Fluent builder constructing a request to AdminDeleteUser.

AdminDeleteUserAttributes

Fluent builder constructing a request to AdminDeleteUserAttributes.

AdminDisableProviderForUser

Fluent builder constructing a request to AdminDisableProviderForUser.

AdminDisableUser

Fluent builder constructing a request to AdminDisableUser.

AdminEnableUser

Fluent builder constructing a request to AdminEnableUser.

AdminForgetDevice

Fluent builder constructing a request to AdminForgetDevice.

AdminGetDevice

Fluent builder constructing a request to AdminGetDevice.

AdminGetUser

Fluent builder constructing a request to AdminGetUser.

AdminInitiateAuth

Fluent builder constructing a request to AdminInitiateAuth.

AdminLinkProviderForUser

Fluent builder constructing a request to AdminLinkProviderForUser.

AdminListDevices

Fluent builder constructing a request to AdminListDevices.

AdminListGroupsForUser

Fluent builder constructing a request to AdminListGroupsForUser.

AdminListUserAuthEvents

Fluent builder constructing a request to AdminListUserAuthEvents.

AdminRemoveUserFromGroup

Fluent builder constructing a request to AdminRemoveUserFromGroup.

AdminResetUserPassword

Fluent builder constructing a request to AdminResetUserPassword.

AdminRespondToAuthChallenge

Fluent builder constructing a request to AdminRespondToAuthChallenge.

AdminSetUserMFAPreference

Fluent builder constructing a request to AdminSetUserMFAPreference.

AdminSetUserPassword

Fluent builder constructing a request to AdminSetUserPassword.

AdminSetUserSettings

Fluent builder constructing a request to AdminSetUserSettings.

AdminUpdateAuthEventFeedback

Fluent builder constructing a request to AdminUpdateAuthEventFeedback.

AdminUpdateDeviceStatus

Fluent builder constructing a request to AdminUpdateDeviceStatus.

AdminUpdateUserAttributes

Fluent builder constructing a request to AdminUpdateUserAttributes.

AdminUserGlobalSignOut

Fluent builder constructing a request to AdminUserGlobalSignOut.

AssociateSoftwareToken

Fluent builder constructing a request to AssociateSoftwareToken.

ChangePassword

Fluent builder constructing a request to ChangePassword.

ConfirmDevice

Fluent builder constructing a request to ConfirmDevice.

ConfirmForgotPassword

Fluent builder constructing a request to ConfirmForgotPassword.

ConfirmSignUp

Fluent builder constructing a request to ConfirmSignUp.

CreateGroup

Fluent builder constructing a request to CreateGroup.

CreateIdentityProvider

Fluent builder constructing a request to CreateIdentityProvider.

CreateResourceServer

Fluent builder constructing a request to CreateResourceServer.

CreateUserImportJob

Fluent builder constructing a request to CreateUserImportJob.

CreateUserPool

Fluent builder constructing a request to CreateUserPool.

CreateUserPoolClient

Fluent builder constructing a request to CreateUserPoolClient.

CreateUserPoolDomain

Fluent builder constructing a request to CreateUserPoolDomain.

DeleteGroup

Fluent builder constructing a request to DeleteGroup.

DeleteIdentityProvider

Fluent builder constructing a request to DeleteIdentityProvider.

DeleteResourceServer

Fluent builder constructing a request to DeleteResourceServer.

DeleteUser

Fluent builder constructing a request to DeleteUser.

DeleteUserAttributes

Fluent builder constructing a request to DeleteUserAttributes.

DeleteUserPool

Fluent builder constructing a request to DeleteUserPool.

DeleteUserPoolClient

Fluent builder constructing a request to DeleteUserPoolClient.

DeleteUserPoolDomain

Fluent builder constructing a request to DeleteUserPoolDomain.

DescribeIdentityProvider

Fluent builder constructing a request to DescribeIdentityProvider.

DescribeResourceServer

Fluent builder constructing a request to DescribeResourceServer.

DescribeRiskConfiguration

Fluent builder constructing a request to DescribeRiskConfiguration.

DescribeUserImportJob

Fluent builder constructing a request to DescribeUserImportJob.

DescribeUserPool

Fluent builder constructing a request to DescribeUserPool.

DescribeUserPoolClient

Fluent builder constructing a request to DescribeUserPoolClient.

DescribeUserPoolDomain

Fluent builder constructing a request to DescribeUserPoolDomain.

ForgetDevice

Fluent builder constructing a request to ForgetDevice.

ForgotPassword

Fluent builder constructing a request to ForgotPassword.

GetCSVHeader

Fluent builder constructing a request to GetCSVHeader.

GetDevice

Fluent builder constructing a request to GetDevice.

GetGroup

Fluent builder constructing a request to GetGroup.

GetIdentityProviderByIdentifier

Fluent builder constructing a request to GetIdentityProviderByIdentifier.

GetSigningCertificate

Fluent builder constructing a request to GetSigningCertificate.

GetUICustomization

Fluent builder constructing a request to GetUICustomization.

GetUser

Fluent builder constructing a request to GetUser.

GetUserAttributeVerificationCode

Fluent builder constructing a request to GetUserAttributeVerificationCode.

GetUserPoolMfaConfig

Fluent builder constructing a request to GetUserPoolMfaConfig.

GlobalSignOut

Fluent builder constructing a request to GlobalSignOut.

InitiateAuth

Fluent builder constructing a request to InitiateAuth.

ListDevices

Fluent builder constructing a request to ListDevices.

ListGroups

Fluent builder constructing a request to ListGroups.

ListIdentityProviders

Fluent builder constructing a request to ListIdentityProviders.

ListResourceServers

Fluent builder constructing a request to ListResourceServers.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ListUserImportJobs

Fluent builder constructing a request to ListUserImportJobs.

ListUserPoolClients

Fluent builder constructing a request to ListUserPoolClients.

ListUserPools

Fluent builder constructing a request to ListUserPools.

ListUsers

Fluent builder constructing a request to ListUsers.

ListUsersInGroup

Fluent builder constructing a request to ListUsersInGroup.

ResendConfirmationCode

Fluent builder constructing a request to ResendConfirmationCode.

RespondToAuthChallenge

Fluent builder constructing a request to RespondToAuthChallenge.

RevokeToken

Fluent builder constructing a request to RevokeToken.

SetRiskConfiguration

Fluent builder constructing a request to SetRiskConfiguration.

SetUICustomization

Fluent builder constructing a request to SetUICustomization.

SetUserMFAPreference

Fluent builder constructing a request to SetUserMFAPreference.

SetUserPoolMfaConfig

Fluent builder constructing a request to SetUserPoolMfaConfig.

SetUserSettings

Fluent builder constructing a request to SetUserSettings.

SignUp

Fluent builder constructing a request to SignUp.

StartUserImportJob

Fluent builder constructing a request to StartUserImportJob.

StopUserImportJob

Fluent builder constructing a request to StopUserImportJob.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateAuthEventFeedback

Fluent builder constructing a request to UpdateAuthEventFeedback.

UpdateDeviceStatus

Fluent builder constructing a request to UpdateDeviceStatus.

UpdateGroup

Fluent builder constructing a request to UpdateGroup.

UpdateIdentityProvider

Fluent builder constructing a request to UpdateIdentityProvider.

UpdateResourceServer

Fluent builder constructing a request to UpdateResourceServer.

UpdateUserAttributes

Fluent builder constructing a request to UpdateUserAttributes.

UpdateUserPool

Fluent builder constructing a request to UpdateUserPool.

UpdateUserPoolClient

Fluent builder constructing a request to UpdateUserPoolClient.

UpdateUserPoolDomain

Fluent builder constructing a request to UpdateUserPoolDomain.

VerifySoftwareToken

Fluent builder constructing a request to VerifySoftwareToken.

VerifyUserAttribute

Fluent builder constructing a request to VerifyUserAttribute.