Struct aws_sdk_cognitoidentityprovider::client::fluent_builders::AdminRespondToAuthChallenge

pub struct AdminRespondToAuthChallenge<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> AdminRespondToAuthChallenge<C>

pub async fn send(
self
) -> Result<AdminRespondToAuthChallengeOutput, SdkError<AdminRespondToAuthChallengeError>> where
C: SmithyConnector,

pub fn user_pool_id(self, input: impl Into<String>) -> Self

The ID of the Amazon Cognito user pool.

pub fn set_user_pool_id(self, input: Option<String>) -> Self

pub fn client_id(self, input: impl Into<String>) -> Self

The app client ID.

pub fn set_client_id(self, input: Option<String>) -> Self

pub fn challenge_name(self, input: ChallengeNameType) -> Self

The challenge name. For more information, see AdminInitiateAuth.

pub fn set_challenge_name(self, input: Option<ChallengeNameType>) -> Self

pub fn challenge_responses(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

The challenge responses. These are inputs corresponding to the value of ChallengeName, for example:

SMS_MFA: SMS_MFA_CODE, USERNAME, SECRET_HASH (if app client is configured with client secret).

PASSWORD_VERIFIER: PASSWORD_CLAIM_SIGNATURE, PASSWORD_CLAIM_SECRET_BLOCK, TIMESTAMP, USERNAME, SECRET_HASH (if app client is configured with client secret).

ADMIN_NO_SRP_AUTH: PASSWORD, USERNAME, SECRET_HASH (if app client is configured with client secret).

NEW_PASSWORD_REQUIRED: NEW_PASSWORD, any other required attributes, USERNAME, SECRET_HASH (if app client is configured with client secret).

MFA_SETUP requires USERNAME, plus you need to use the session value returned by VerifySoftwareToken in the Session parameter.

The value of the USERNAME attribute must be the user's actual username, not an alias (such as email address or phone number). To make this easier, the AdminInitiateAuth response includes the actual username value in the USERNAMEUSER_ID_FOR_SRP attribute, even if you specified an alias in your call to AdminInitiateAuth.

pub fn set_challenge_responses(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn session(self, input: impl Into<String>) -> Self

The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

pub fn set_session(self, input: Option<String>) -> Self

pub fn analytics_metadata(self, input: AnalyticsMetadataType) -> Self

The analytics metadata for collecting Amazon Pinpoint metrics for AdminRespondToAuthChallenge calls.

pub fn set_analytics_metadata(
self,
input: Option<AnalyticsMetadataType>
) -> Self

pub fn context_data(self, input: ContextDataType) -> Self

Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

pub fn set_context_data(self, input: Option<ContextDataType>) -> Self

pub fn client_metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

A map of custom key-value pairs that you can provide as input for any custom workflows that this action triggers.

You create custom workflows by assigning AWS Lambda functions to user pool triggers. When you use the AdminRespondToAuthChallenge API action, Amazon Cognito invokes any functions that are assigned to the following triggers: pre sign-up, custom message, post authentication, user migration, pre token generation, define auth challenge, create auth challenge, and verify auth challenge response. When Amazon Cognito invokes any of these functions, it passes a JSON payload, which the function receives as input. This payload contains a clientMetadata attribute, which provides the data that you assigned to the ClientMetadata parameter in your AdminRespondToAuthChallenge request. In your function code in AWS Lambda, you can process the clientMetadata value to enhance your workflow for your specific needs.

For more information, see Customizing User Pool Workflows with Lambda Triggers in the Amazon Cognito Developer Guide.

Take the following limitations into consideration when you use the ClientMetadata parameter:

Amazon Cognito does not store the ClientMetadata value. This data is available only to AWS Lambda triggers that are assigned to a user pool to support custom workflows. If your user pool configuration does not include triggers, the ClientMetadata parameter serves no purpose.

Amazon Cognito does not validate the ClientMetadata value.

Amazon Cognito does not encrypt the the ClientMetadata value, so don't use it to provide sensitive information.

pub fn set_client_metadata(self, input: Option<HashMap<String, String>>) -> Self