Struct aws_sdk_cognitoidentityprovider::client::fluent_builders::CreateUserPool

pub struct CreateUserPool<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateUserPool<C>

pub async fn send(
self
) -> Result<CreateUserPoolOutput, SdkError<CreateUserPoolError>> where
C: SmithyConnector,

pub fn pool_name(self, input: impl Into<String>) -> Self

A string used to name the user pool.

pub fn set_pool_name(self, input: Option<String>) -> Self

pub fn policies(self, input: UserPoolPolicyType) -> Self

The policies associated with the new user pool.

pub fn set_policies(self, input: Option<UserPoolPolicyType>) -> Self

pub fn lambda_config(self, input: LambdaConfigType) -> Self

The Lambda trigger configuration information for the new user pool.

In a push model, event sources (such as Amazon S3 and custom applications) need permission to invoke a function. So you will need to make an extra call to add permission for these event sources to invoke your Lambda function.

For more information on using the Lambda API to add permission, see AddPermission .

For adding permission using the AWS CLI, see add-permission .

pub fn set_lambda_config(self, input: Option<LambdaConfigType>) -> Self

pub fn auto_verified_attributes(
self,
inp: impl Into<VerifiedAttributeType>
) -> Self

The attributes to be auto-verified. Possible values: email, phone_number.

pub fn set_auto_verified_attributes(
self,
input: Option<Vec<VerifiedAttributeType>>
) -> Self

pub fn alias_attributes(self, inp: impl Into<AliasAttributeType>) -> Self

Attributes supported as an alias for this user pool. Possible values: phone_number, email, or preferred_username.

pub fn set_alias_attributes(
self,
input: Option<Vec<AliasAttributeType>>
) -> Self

pub fn username_attributes(self, inp: impl Into<UsernameAttributeType>) -> Self

Specifies whether email addresses or phone numbers can be specified as usernames when a user signs up.

pub fn set_username_attributes(
self,
input: Option<Vec<UsernameAttributeType>>
) -> Self

pub fn sms_verification_message(self, input: impl Into<String>) -> Self

A string representing the SMS verification message.

pub fn set_sms_verification_message(self, input: Option<String>) -> Self

pub fn email_verification_message(self, input: impl Into<String>) -> Self

A string representing the email verification message. EmailVerificationMessage is allowed only if EmailSendingAccount is DEVELOPER.

pub fn set_email_verification_message(self, input: Option<String>) -> Self

pub fn email_verification_subject(self, input: impl Into<String>) -> Self

A string representing the email verification subject. EmailVerificationSubject is allowed only if EmailSendingAccount is DEVELOPER.

pub fn set_email_verification_subject(self, input: Option<String>) -> Self

pub fn verification_message_template(
self,
input: VerificationMessageTemplateType
) -> Self

The template for the verification message that the user sees when the app requests permission to access the user's information.

pub fn set_verification_message_template(
self,
input: Option<VerificationMessageTemplateType>
) -> Self

pub fn sms_authentication_message(self, input: impl Into<String>) -> Self

A string representing the SMS authentication message.

pub fn set_sms_authentication_message(self, input: Option<String>) -> Self

pub fn mfa_configuration(self, input: UserPoolMfaType) -> Self

Specifies MFA configuration details.

pub fn set_mfa_configuration(self, input: Option<UserPoolMfaType>) -> Self

pub fn device_configuration(self, input: DeviceConfigurationType) -> Self

The device configuration.

pub fn set_device_configuration(
self,
input: Option<DeviceConfigurationType>
) -> Self

pub fn email_configuration(self, input: EmailConfigurationType) -> Self

The email configuration.

pub fn set_email_configuration(
self,
input: Option<EmailConfigurationType>
) -> Self

pub fn sms_configuration(self, input: SmsConfigurationType) -> Self

The SMS configuration.

pub fn set_sms_configuration(self, input: Option<SmsConfigurationType>) -> Self

pub fn user_pool_tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

The tag keys and values to assign to the user pool. A tag is a label that you can use to categorize and manage user pools in different ways, such as by purpose, owner, environment, or other criteria.

pub fn set_user_pool_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn admin_create_user_config(self, input: AdminCreateUserConfigType) -> Self

The configuration for AdminCreateUser requests.

pub fn set_admin_create_user_config(
self,
input: Option<AdminCreateUserConfigType>
) -> Self

pub fn schema(self, inp: impl Into<SchemaAttributeType>) -> Self

An array of schema attributes for the new user pool. These attributes can be standard or custom attributes.

pub fn set_schema(self, input: Option<Vec<SchemaAttributeType>>) -> Self

pub fn user_pool_add_ons(self, input: UserPoolAddOnsType) -> Self

Used to enable advanced security risk detection. Set the key AdvancedSecurityMode to the value "AUDIT".

pub fn set_user_pool_add_ons(self, input: Option<UserPoolAddOnsType>) -> Self

pub fn username_configuration(self, input: UsernameConfigurationType) -> Self

You can choose to set case sensitivity on the username input for the selected sign-in option. For example, when this is set to False, users will be able to sign in using either "username" or "Username". This configuration is immutable once it has been set. For more information, see UsernameConfigurationType.

pub fn set_username_configuration(
self,
input: Option<UsernameConfigurationType>
) -> Self

pub fn account_recovery_setting(self, input: AccountRecoverySettingType) -> Self

Use this setting to define which verified available method a user can use to recover their password when they call ForgotPassword. It allows you to define a preferred method when a user has more than one method available. With this setting, SMS does not qualify for a valid password recovery mechanism if the user also has SMS MFA enabled. In the absence of this setting, Cognito uses the legacy behavior to determine the recovery method where SMS is preferred over email.

pub fn set_account_recovery_setting(
self,
input: Option<AccountRecoverySettingType>
) -> Self