Struct aws_sdk_cognitoidentityprovider::client::fluent_builders::CreateUserPoolClient

pub struct CreateUserPoolClient<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateUserPoolClient<C>

pub async fn send(
self
) -> Result<CreateUserPoolClientOutput, SdkError<CreateUserPoolClientError>> where
C: SmithyConnector,

pub fn user_pool_id(self, input: impl Into<String>) -> Self

The user pool ID for the user pool where you want to create a user pool client.

pub fn set_user_pool_id(self, input: Option<String>) -> Self

pub fn client_name(self, input: impl Into<String>) -> Self

The client name for the user pool client you would like to create.

pub fn set_client_name(self, input: Option<String>) -> Self

pub fn generate_secret(self, input: bool) -> Self

Boolean to specify whether you want to generate a secret for the user pool client being created.

pub fn set_generate_secret(self, input: Option<bool>) -> Self

pub fn refresh_token_validity(self, input: i32) -> Self

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

pub fn set_refresh_token_validity(self, input: Option<i32>) -> Self

pub fn access_token_validity(self, input: i32) -> Self

The time limit, between 5 minutes and 1 day, after which the access token is no longer valid and cannot be used. This value will be overridden if you have entered a value in TokenValidityUnits.

pub fn set_access_token_validity(self, input: Option<i32>) -> Self

pub fn id_token_validity(self, input: i32) -> Self

The time limit, between 5 minutes and 1 day, after which the ID token is no longer valid and cannot be used. This value will be overridden if you have entered a value in TokenValidityUnits.

pub fn set_id_token_validity(self, input: Option<i32>) -> Self

pub fn token_validity_units(self, input: TokenValidityUnitsType) -> Self

The units in which the validity times are represented in. Default for RefreshToken is days, and default for ID and access tokens are hours.

pub fn set_token_validity_units(
self,
input: Option<TokenValidityUnitsType>
) -> Self

pub fn read_attributes(self, inp: impl Into<String>) -> Self

The read attributes.

pub fn set_read_attributes(self, input: Option<Vec<String>>) -> Self

pub fn write_attributes(self, inp: impl Into<String>) -> Self

The user pool attributes that the app client can write to.

If your app client allows users to sign in through an identity provider, this array must include all attributes that are mapped to identity provider attributes. Amazon Cognito updates mapped attributes when users sign in to your application through an identity provider. If your app client lacks write access to a mapped attribute, Amazon Cognito throws an error when it attempts to update the attribute. For more information, see Specifying Identity Provider Attribute Mappings for Your User Pool.

pub fn set_write_attributes(self, input: Option<Vec<String>>) -> Self

pub fn explicit_auth_flows(self, inp: impl Into<ExplicitAuthFlowsType>) -> Self

The authentication flows that are supported by the user pool clients. Flow names without the ALLOW_ prefix are deprecated in favor of new names with the ALLOW_ prefix. Note that values with ALLOW_ prefix cannot be used along with values without ALLOW_ prefix.

Valid values include:

ALLOW_ADMIN_USER_PASSWORD_AUTH: Enable admin based user password authentication flow ADMIN_USER_PASSWORD_AUTH. This setting replaces the ADMIN_NO_SRP_AUTH setting. With this authentication flow, Cognito receives the password in the request instead of using the SRP (Secure Remote Password protocol) protocol to verify passwords.

ALLOW_CUSTOM_AUTH: Enable Lambda trigger based authentication.

ALLOW_USER_PASSWORD_AUTH: Enable user password-based authentication. In this flow, Cognito receives the password in the request instead of using the SRP protocol to verify passwords.

ALLOW_USER_SRP_AUTH: Enable SRP based authentication.

ALLOW_REFRESH_TOKEN_AUTH: Enable authflow to refresh tokens.

pub fn set_explicit_auth_flows(
self,
input: Option<Vec<ExplicitAuthFlowsType>>
) -> Self

pub fn supported_identity_providers(self, inp: impl Into<String>) -> Self

A list of provider names for the identity providers that are supported on this client. The following are supported: COGNITO, Facebook, Google and LoginWithAmazon.

pub fn set_supported_identity_providers(
self,
input: Option<Vec<String>>
) -> Self

pub fn callback_ur_ls(self, inp: impl Into<String>) -> Self

A list of allowed redirect (callback) URLs for the identity providers.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

pub fn set_callback_ur_ls(self, input: Option<Vec<String>>) -> Self

pub fn logout_ur_ls(self, inp: impl Into<String>) -> Self

A list of allowed logout URLs for the identity providers.

pub fn set_logout_ur_ls(self, input: Option<Vec<String>>) -> Self

pub fn default_redirect_uri(self, input: impl Into<String>) -> Self

The default redirect URI. Must be in the CallbackURLs list.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

pub fn set_default_redirect_uri(self, input: Option<String>) -> Self

pub fn allowed_o_auth_flows(self, inp: impl Into<OAuthFlowType>) -> Self

The allowed OAuth flows.

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint.

Set to implicit to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

Set to client_credentials to specify that the client should get the access token (and, optionally, ID token, based on scopes) from the token endpoint using a combination of client and client_secret.

pub fn set_allowed_o_auth_flows(self, input: Option<Vec<OAuthFlowType>>) -> Self

pub fn allowed_o_auth_scopes(self, inp: impl Into<String>) -> Self

The allowed OAuth scopes. Possible values provided by OAuth are: phone, email, openid, and profile. Possible values provided by AWS are: aws.cognito.signin.user.admin. Custom scopes created in Resource Servers are also supported.

pub fn set_allowed_o_auth_scopes(self, input: Option<Vec<String>>) -> Self

pub fn allowed_o_auth_flows_user_pool_client(self, input: bool) -> Self

Set to true if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

pub fn set_allowed_o_auth_flows_user_pool_client(
self,
input: Option<bool>
) -> Self

pub fn analytics_configuration(self, input: AnalyticsConfigurationType) -> Self

The Amazon Pinpoint analytics configuration for collecting metrics for this user pool.

In regions where Pinpoint is not available, Cognito User Pools only supports sending events to Amazon Pinpoint projects in us-east-1. In regions where Pinpoint is available, Cognito User Pools will support sending events to Amazon Pinpoint projects within that same region.

pub fn set_analytics_configuration(
self,
input: Option<AnalyticsConfigurationType>
) -> Self

pub fn prevent_user_existence_errors(
self,
input: PreventUserExistenceErrorTypes
) -> Self

Use this setting to choose which errors and responses are returned by Cognito APIs during authentication, account confirmation, and password recovery when the user does not exist in the user pool. When set to ENABLED and the user does not exist, authentication returns an error indicating either the username or password was incorrect, and account confirmation and password recovery return a response indicating a code was sent to a simulated destination. When set to LEGACY, those APIs will return a UserNotFoundException exception if the user does not exist in the user pool.

Valid values include:

ENABLED - This prevents user existence-related errors.

LEGACY - This represents the old behavior of Cognito where user existence related errors are not prevented.

After February 15th 2020, the value of PreventUserExistenceErrors will default to ENABLED for newly created user pool clients if no value is provided.

pub fn set_prevent_user_existence_errors(
self,
input: Option<PreventUserExistenceErrorTypes>
) -> Self

pub fn enable_token_revocation(self, input: bool) -> Self

Enables or disables token revocation. For more information about revoking tokens, see RevokeToken.

If you don't include this parameter, token revocation is automatically enabled for the new user pool client.

pub fn set_enable_token_revocation(self, input: Option<bool>) -> Self