Struct aws_sdk_cognitoidentityprovider::client::fluent_builders::UpdateUserPool

pub struct UpdateUserPool<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateUserPool<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateUserPoolOutput, SdkError<UpdateUserPoolError>> where
R::Policy: SmithyRetryPolicy<UpdateUserPoolInputOperationOutputAlias, UpdateUserPoolOutput, UpdateUserPoolError, UpdateUserPoolInputOperationRetryAlias>,

pub fn user_pool_id(self, inp: impl Into<String>) -> Self

The user pool ID for the user pool you want to update.

pub fn set_user_pool_id(self, input: Option<String>) -> Self

pub fn policies(self, inp: UserPoolPolicyType) -> Self

A container with the policies you wish to update in a user pool.

pub fn set_policies(self, input: Option<UserPoolPolicyType>) -> Self

pub fn lambda_config(self, inp: LambdaConfigType) -> Self

The Lambda configuration information from the request to update the user pool.

pub fn set_lambda_config(self, input: Option<LambdaConfigType>) -> Self

pub fn auto_verified_attributes(
self,
inp: impl Into<VerifiedAttributeType>
) -> Self

Appends an item to AutoVerifiedAttributes.

To override the contents of this collection use set_auto_verified_attributes.

The attributes that are automatically verified when the Amazon Cognito service makes a request to update user pools.

pub fn set_auto_verified_attributes(
self,
input: Option<Vec<VerifiedAttributeType>>
) -> Self

pub fn sms_verification_message(self, inp: impl Into<String>) -> Self

A container with information about the SMS verification message.

pub fn set_sms_verification_message(self, input: Option<String>) -> Self

pub fn email_verification_message(self, inp: impl Into<String>) -> Self

The contents of the email verification message.

pub fn set_email_verification_message(self, input: Option<String>) -> Self

pub fn email_verification_subject(self, inp: impl Into<String>) -> Self

The subject of the email verification message.

pub fn set_email_verification_subject(self, input: Option<String>) -> Self

pub fn verification_message_template(
self,
inp: VerificationMessageTemplateType
) -> Self

The template for verification messages.

pub fn set_verification_message_template(
self,
input: Option<VerificationMessageTemplateType>
) -> Self

pub fn sms_authentication_message(self, inp: impl Into<String>) -> Self

The contents of the SMS authentication message.

pub fn set_sms_authentication_message(self, input: Option<String>) -> Self

pub fn mfa_configuration(self, inp: UserPoolMfaType) -> Self

Can be one of the following values:

OFF - MFA tokens are not required and cannot be specified during user registration.

ON - MFA tokens are required for all user registrations. You can only specify ON when you are initially creating a user pool. You can use the SetUserPoolMfaConfig API operation to turn MFA "ON" for existing user pools.

OPTIONAL - Users have the option when registering to create an MFA token.

pub fn set_mfa_configuration(self, input: Option<UserPoolMfaType>) -> Self

pub fn device_configuration(self, inp: DeviceConfigurationType) -> Self

Device configuration.

pub fn set_device_configuration(
self,
input: Option<DeviceConfigurationType>
) -> Self

pub fn email_configuration(self, inp: EmailConfigurationType) -> Self

Email configuration.

pub fn set_email_configuration(
self,
input: Option<EmailConfigurationType>
) -> Self

pub fn sms_configuration(self, inp: SmsConfigurationType) -> Self

SMS configuration.

pub fn set_sms_configuration(self, input: Option<SmsConfigurationType>) -> Self

pub fn user_pool_tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to UserPoolTags.

To override the contents of this collection use set_user_pool_tags.

The tag keys and values to assign to the user pool. A tag is a label that you can use to categorize and manage user pools in different ways, such as by purpose, owner, environment, or other criteria.

pub fn set_user_pool_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn admin_create_user_config(self, inp: AdminCreateUserConfigType) -> Self

The configuration for AdminCreateUser requests.

pub fn set_admin_create_user_config(
self,
input: Option<AdminCreateUserConfigType>
) -> Self

pub fn user_pool_add_ons(self, inp: UserPoolAddOnsType) -> Self

Used to enable advanced security risk detection. Set the key AdvancedSecurityMode to the value "AUDIT".

pub fn set_user_pool_add_ons(self, input: Option<UserPoolAddOnsType>) -> Self

pub fn account_recovery_setting(self, inp: AccountRecoverySettingType) -> Self

Use this setting to define which verified available method a user can use to recover their password when they call ForgotPassword. It allows you to define a preferred method when a user has more than one method available. With this setting, SMS does not qualify for a valid password recovery mechanism if the user also has SMS MFA enabled. In the absence of this setting, Cognito uses the legacy behavior to determine the recovery method where SMS is preferred over email.

pub fn set_account_recovery_setting(
self,
input: Option<AccountRecoverySettingType>
) -> Self