Enum aws_sdk_cognitoidentityprovider::SdkError

pub enum SdkError<E> {
    ConstructionFailure(Box<dyn Error + Sync + Send + 'static, Global>),
    DispatchFailure(Box<dyn Error + Sync + Send + 'static, Global>),
    ResponseError {
        raw: Response<SdkBody>,
        err: Box<dyn Error + Sync + Send + 'static, Global>,
    },
    ServiceError {
        err: E,
        raw: Response<SdkBody>,
    },
}

Failing Sdk Result

Variants

ConstructionFailure(Box<dyn Error + Sync + Send + 'static, Global>)

The request failed during construction. It was not dispatched over the network.

DispatchFailure(Box<dyn Error + Sync + Send + 'static, Global>)

The request failed during dispatch. An HTTP response was not received. The request MAY have been sent.

ResponseError

A response was received but it was not parseable according the the protocol (for example the server hung up while the body was being read)

Fields of ResponseError

raw: Response<SdkBody>err: Box<dyn Error + Sync + Send + 'static, Global>

ServiceError

An error response was received from the service

Fields of ServiceError

err: Eraw: Response<SdkBody>