Module aws_sdk_cognitoidentityprovider::error

Modules

alias_exists_exception

See AliasExistsException

code_delivery_failure_exception

See CodeDeliveryFailureException

code_mismatch_exception

See CodeMismatchException

concurrent_modification_exception

See ConcurrentModificationException

duplicate_provider_exception

See DuplicateProviderException

enable_software_token_mfa_exception

See EnableSoftwareTokenMfaException

expired_code_exception

See ExpiredCodeException

group_exists_exception

See GroupExistsException

internal_error_exception

See InternalErrorException

invalid_email_role_access_policy_exception

See InvalidEmailRoleAccessPolicyException

invalid_lambda_response_exception

See InvalidLambdaResponseException

invalid_o_auth_flow_exception

See InvalidOAuthFlowException

invalid_parameter_exception

See InvalidParameterException

invalid_password_exception

See InvalidPasswordException

invalid_sms_role_access_policy_exception

See InvalidSmsRoleAccessPolicyException

invalid_sms_role_trust_relationship_exception

See InvalidSmsRoleTrustRelationshipException

invalid_user_pool_configuration_exception

See InvalidUserPoolConfigurationException

limit_exceeded_exception

See LimitExceededException

mfa_method_not_found_exception

See MfaMethodNotFoundException

not_authorized_exception

See NotAuthorizedException

password_reset_required_exception

See PasswordResetRequiredException

precondition_not_met_exception

See PreconditionNotMetException

resource_not_found_exception

See ResourceNotFoundException

scope_does_not_exist_exception

See ScopeDoesNotExistException

software_token_mfa_not_found_exception

See SoftwareTokenMfaNotFoundException

too_many_failed_attempts_exception

See TooManyFailedAttemptsException

too_many_requests_exception

See TooManyRequestsException

unauthorized_exception

See UnauthorizedException

unexpected_lambda_exception

See UnexpectedLambdaException

unsupported_identity_provider_exception

See UnsupportedIdentityProviderException

unsupported_operation_exception

See UnsupportedOperationException

unsupported_token_type_exception

See UnsupportedTokenTypeException

unsupported_user_state_exception

See UnsupportedUserStateException

user_import_in_progress_exception

See UserImportInProgressException

user_lambda_validation_exception

See UserLambdaValidationException

user_not_confirmed_exception

See UserNotConfirmedException

user_not_found_exception

See UserNotFoundException

user_pool_add_on_not_enabled_exception

See UserPoolAddOnNotEnabledException

user_pool_tagging_exception

See UserPoolTaggingException

username_exists_exception

See UsernameExistsException

Structs

AddCustomAttributesError

AdminAddUserToGroupError

AdminConfirmSignUpError

AdminCreateUserError

AdminDeleteUserAttributesError

AdminDeleteUserError

AdminDisableProviderForUserError

AdminDisableUserError

AdminEnableUserError

AdminForgetDeviceError

AdminGetDeviceError

AdminGetUserError

AdminInitiateAuthError

AdminLinkProviderForUserError

AdminListDevicesError

AdminListGroupsForUserError

AdminListUserAuthEventsError

AdminRemoveUserFromGroupError

AdminResetUserPasswordError

AdminRespondToAuthChallengeError

AdminSetUserMFAPreferenceError

AdminSetUserPasswordError

AdminSetUserSettingsError

AdminUpdateAuthEventFeedbackError

AdminUpdateDeviceStatusError

AdminUpdateUserAttributesError

AdminUserGlobalSignOutError

AliasExistsException

This exception is thrown when a user tries to confirm the account with an email or

AssociateSoftwareTokenError

ChangePasswordError

CodeDeliveryFailureException

This exception is thrown when a verification code fails to deliver

CodeMismatchException

This exception is thrown if the provided code does not match what the server was

ConcurrentModificationException

This exception is thrown if two or more modifications are happening

ConfirmDeviceError

ConfirmForgotPasswordError

ConfirmSignUpError

CreateGroupError

CreateIdentityProviderError

CreateResourceServerError

CreateUserImportJobError

CreateUserPoolClientError

CreateUserPoolDomainError

CreateUserPoolError

DeleteGroupError

DeleteIdentityProviderError

DeleteResourceServerError

DeleteUserAttributesError

DeleteUserError

DeleteUserPoolClientError

DeleteUserPoolDomainError

DeleteUserPoolError

DescribeIdentityProviderError

DescribeResourceServerError

DescribeRiskConfigurationError

DescribeUserImportJobError

DescribeUserPoolClientError

DescribeUserPoolDomainError

DescribeUserPoolError

DuplicateProviderException

This exception is thrown when the provider is already supported by the user

EnableSoftwareTokenMfaException

This exception is thrown when there is a code mismatch and the service fails to

ExpiredCodeException

This exception is thrown if a code has expired.

ForgetDeviceError

ForgotPasswordError

GetCSVHeaderError

GetDeviceError

GetGroupError

GetIdentityProviderByIdentifierError

GetSigningCertificateError

GetUICustomizationError

GetUserAttributeVerificationCodeError

GetUserError

GetUserPoolMfaConfigError

GlobalSignOutError

GroupExistsException

This exception is thrown when Amazon Cognito encounters a group that already exists in

InitiateAuthError

InternalErrorException

This exception is thrown when Amazon Cognito encounters an internal error.

InvalidEmailRoleAccessPolicyException

This exception is thrown when Amazon Cognito is not allowed to use your email

InvalidLambdaResponseException

This exception is thrown when the Amazon Cognito service encounters an invalid AWS

InvalidOAuthFlowException

This exception is thrown when the specified OAuth flow is invalid.

InvalidParameterException

This exception is thrown when the Amazon Cognito service encounters an invalid

InvalidPasswordException

This exception is thrown when the Amazon Cognito service encounters an invalid

InvalidSmsRoleAccessPolicyException

This exception is returned when the role provided for SMS configuration does not have

InvalidSmsRoleTrustRelationshipException

This exception is thrown when the trust relationship is invalid for the role provided

InvalidUserPoolConfigurationException

This exception is thrown when the user pool configuration is invalid.

LimitExceededException

This exception is thrown when a user exceeds the limit for a requested AWS

ListDevicesError

ListGroupsError

ListIdentityProvidersError

ListResourceServersError

ListTagsForResourceError

ListUserImportJobsError

ListUserPoolClientsError

ListUserPoolsError

ListUsersError

ListUsersInGroupError

MfaMethodNotFoundException

This exception is thrown when Amazon Cognito cannot find a multi-factor authentication

NotAuthorizedException

This exception is thrown when a user is not authorized.

PasswordResetRequiredException

This exception is thrown when a password reset is required.

PreconditionNotMetException

This exception is thrown when a precondition is not met.

ResendConfirmationCodeError

ResourceNotFoundException

This exception is thrown when the Amazon Cognito service cannot find the requested

RespondToAuthChallengeError

RevokeTokenError

ScopeDoesNotExistException

This exception is thrown when the specified scope does not exist.

SetRiskConfigurationError

SetUICustomizationError

SetUserMFAPreferenceError

SetUserPoolMfaConfigError

SetUserSettingsError

SignUpError

SoftwareTokenMfaNotFoundException

This exception is thrown when the software token TOTP multi-factor authentication

StartUserImportJobError

StopUserImportJobError

TagResourceError

TooManyFailedAttemptsException

This exception is thrown when the user has made too many failed attempts for a given

TooManyRequestsException

This exception is thrown when the user has made too many requests for a given

UnauthorizedException

This exception is thrown when the request is not authorized. This can happen due to an invalid access token in the request.

UnexpectedLambdaException

This exception is thrown when the Amazon Cognito service encounters an unexpected

UnsupportedIdentityProviderException

This exception is thrown when the specified identifier is not supported.

UnsupportedOperationException

This exception is thrown when you attempt to perform an operation that is not enabled for the user pool client.

UnsupportedTokenTypeException

This exception is thrown when an unsupported token is passed to an operation.

UnsupportedUserStateException

The request failed because the user is in an unsupported state.

UntagResourceError

UpdateAuthEventFeedbackError

UpdateDeviceStatusError

UpdateGroupError

UpdateIdentityProviderError

UpdateResourceServerError

UpdateUserAttributesError

UpdateUserPoolClientError

UpdateUserPoolDomainError

UpdateUserPoolError

UserImportInProgressException

This exception is thrown when you are trying to modify a user pool while a user import

UserLambdaValidationException

This exception is thrown when the Amazon Cognito service encounters a user validation

UserNotConfirmedException

This exception is thrown when a user is not confirmed successfully.

UserNotFoundException

This exception is thrown when a user is not found.

UserPoolAddOnNotEnabledException

This exception is thrown when user pool add-ons are not enabled.

UserPoolTaggingException

This exception is thrown when a user pool tag cannot be set or updated.

UsernameExistsException

This exception is thrown when Amazon Cognito encounters a user name that already

VerifySoftwareTokenError

VerifyUserAttributeError

Enums

AddCustomAttributesErrorKind

AdminAddUserToGroupErrorKind

AdminConfirmSignUpErrorKind

AdminCreateUserErrorKind

AdminDeleteUserAttributesErrorKind

AdminDeleteUserErrorKind

AdminDisableProviderForUserErrorKind

AdminDisableUserErrorKind

AdminEnableUserErrorKind

AdminForgetDeviceErrorKind

AdminGetDeviceErrorKind

AdminGetUserErrorKind

AdminInitiateAuthErrorKind

AdminLinkProviderForUserErrorKind

AdminListDevicesErrorKind

AdminListGroupsForUserErrorKind

AdminListUserAuthEventsErrorKind

AdminRemoveUserFromGroupErrorKind

AdminResetUserPasswordErrorKind

AdminRespondToAuthChallengeErrorKind

AdminSetUserMFAPreferenceErrorKind

AdminSetUserPasswordErrorKind

AdminSetUserSettingsErrorKind

AdminUpdateAuthEventFeedbackErrorKind

AdminUpdateDeviceStatusErrorKind

AdminUpdateUserAttributesErrorKind

AdminUserGlobalSignOutErrorKind

AssociateSoftwareTokenErrorKind

ChangePasswordErrorKind

ConfirmDeviceErrorKind

ConfirmForgotPasswordErrorKind

ConfirmSignUpErrorKind

CreateGroupErrorKind

CreateIdentityProviderErrorKind

CreateResourceServerErrorKind

CreateUserImportJobErrorKind

CreateUserPoolClientErrorKind

CreateUserPoolDomainErrorKind

CreateUserPoolErrorKind

DeleteGroupErrorKind

DeleteIdentityProviderErrorKind

DeleteResourceServerErrorKind

DeleteUserAttributesErrorKind

DeleteUserErrorKind

DeleteUserPoolClientErrorKind

DeleteUserPoolDomainErrorKind

DeleteUserPoolErrorKind

DescribeIdentityProviderErrorKind

DescribeResourceServerErrorKind

DescribeRiskConfigurationErrorKind

DescribeUserImportJobErrorKind

DescribeUserPoolClientErrorKind

DescribeUserPoolDomainErrorKind

DescribeUserPoolErrorKind

ForgetDeviceErrorKind

ForgotPasswordErrorKind

GetCSVHeaderErrorKind

GetDeviceErrorKind

GetGroupErrorKind

GetIdentityProviderByIdentifierErrorKind

GetSigningCertificateErrorKind

GetUICustomizationErrorKind

GetUserAttributeVerificationCodeErrorKind

GetUserErrorKind

GetUserPoolMfaConfigErrorKind

GlobalSignOutErrorKind

InitiateAuthErrorKind

ListDevicesErrorKind

ListGroupsErrorKind

ListIdentityProvidersErrorKind

ListResourceServersErrorKind

ListTagsForResourceErrorKind

ListUserImportJobsErrorKind

ListUserPoolClientsErrorKind

ListUserPoolsErrorKind

ListUsersErrorKind

ListUsersInGroupErrorKind

ResendConfirmationCodeErrorKind

RespondToAuthChallengeErrorKind

RevokeTokenErrorKind

SetRiskConfigurationErrorKind

SetUICustomizationErrorKind

SetUserMFAPreferenceErrorKind

SetUserPoolMfaConfigErrorKind

SetUserSettingsErrorKind

SignUpErrorKind

StartUserImportJobErrorKind

StopUserImportJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAuthEventFeedbackErrorKind

UpdateDeviceStatusErrorKind

UpdateGroupErrorKind

UpdateIdentityProviderErrorKind

UpdateResourceServerErrorKind

UpdateUserAttributesErrorKind

UpdateUserPoolClientErrorKind

UpdateUserPoolDomainErrorKind

UpdateUserPoolErrorKind

VerifySoftwareTokenErrorKind

VerifyUserAttributeErrorKind