Module aws_sdk_cognitoidentityprovider::input

Modules

add_custom_attributes_input

See AddCustomAttributesInput

admin_add_user_to_group_input

See AdminAddUserToGroupInput

admin_confirm_sign_up_input

See AdminConfirmSignUpInput

admin_create_user_input

See AdminCreateUserInput

admin_delete_user_attributes_input

See AdminDeleteUserAttributesInput

admin_delete_user_input

See AdminDeleteUserInput

admin_disable_provider_for_user_input

See AdminDisableProviderForUserInput

admin_disable_user_input

See AdminDisableUserInput

admin_enable_user_input

See AdminEnableUserInput

admin_forget_device_input

See AdminForgetDeviceInput

admin_get_device_input

See AdminGetDeviceInput

admin_get_user_input

See AdminGetUserInput

admin_initiate_auth_input

See AdminInitiateAuthInput

admin_link_provider_for_user_input

See AdminLinkProviderForUserInput

admin_list_devices_input

See AdminListDevicesInput

admin_list_groups_for_user_input

See AdminListGroupsForUserInput

admin_list_user_auth_events_input

See AdminListUserAuthEventsInput

admin_remove_user_from_group_input

See AdminRemoveUserFromGroupInput

admin_reset_user_password_input

See AdminResetUserPasswordInput

admin_respond_to_auth_challenge_input

See AdminRespondToAuthChallengeInput

admin_set_user_mfa_preference_input

See AdminSetUserMfaPreferenceInput

admin_set_user_password_input

See AdminSetUserPasswordInput

admin_set_user_settings_input

See AdminSetUserSettingsInput

admin_update_auth_event_feedback_input

See AdminUpdateAuthEventFeedbackInput

admin_update_device_status_input

See AdminUpdateDeviceStatusInput

admin_update_user_attributes_input

See AdminUpdateUserAttributesInput

admin_user_global_sign_out_input

See AdminUserGlobalSignOutInput

associate_software_token_input

See AssociateSoftwareTokenInput

change_password_input

See ChangePasswordInput

confirm_device_input

See ConfirmDeviceInput

confirm_forgot_password_input

See ConfirmForgotPasswordInput

confirm_sign_up_input

See ConfirmSignUpInput

create_group_input

See CreateGroupInput

create_identity_provider_input

See CreateIdentityProviderInput

create_resource_server_input

See CreateResourceServerInput

create_user_import_job_input

See CreateUserImportJobInput

create_user_pool_client_input

See CreateUserPoolClientInput

create_user_pool_domain_input

See CreateUserPoolDomainInput

create_user_pool_input

See CreateUserPoolInput

delete_group_input

See DeleteGroupInput

delete_identity_provider_input

See DeleteIdentityProviderInput

delete_resource_server_input

See DeleteResourceServerInput

delete_user_attributes_input

See DeleteUserAttributesInput

delete_user_input

See DeleteUserInput

delete_user_pool_client_input

See DeleteUserPoolClientInput

delete_user_pool_domain_input

See DeleteUserPoolDomainInput

delete_user_pool_input

See DeleteUserPoolInput

describe_identity_provider_input

See DescribeIdentityProviderInput

describe_resource_server_input

See DescribeResourceServerInput

describe_risk_configuration_input

See DescribeRiskConfigurationInput

describe_user_import_job_input

See DescribeUserImportJobInput

describe_user_pool_client_input

See DescribeUserPoolClientInput

describe_user_pool_domain_input

See DescribeUserPoolDomainInput

describe_user_pool_input

See DescribeUserPoolInput

forget_device_input

See ForgetDeviceInput

forgot_password_input

See ForgotPasswordInput

get_csv_header_input

See GetCsvHeaderInput

get_device_input

See GetDeviceInput

get_group_input

See GetGroupInput

get_identity_provider_by_identifier_input

See GetIdentityProviderByIdentifierInput

get_signing_certificate_input

See GetSigningCertificateInput

get_ui_customization_input

See GetUiCustomizationInput

get_user_attribute_verification_code_input

See GetUserAttributeVerificationCodeInput

get_user_input

See GetUserInput

get_user_pool_mfa_config_input

See GetUserPoolMfaConfigInput

global_sign_out_input

See GlobalSignOutInput

initiate_auth_input

See InitiateAuthInput

list_devices_input

See ListDevicesInput

list_groups_input

See ListGroupsInput

list_identity_providers_input

See ListIdentityProvidersInput

list_resource_servers_input

See ListResourceServersInput

list_tags_for_resource_input

See ListTagsForResourceInput

list_user_import_jobs_input

See ListUserImportJobsInput

list_user_pool_clients_input

See ListUserPoolClientsInput

list_user_pools_input

See ListUserPoolsInput

list_users_in_group_input

See ListUsersInGroupInput

list_users_input

See ListUsersInput

resend_confirmation_code_input

See ResendConfirmationCodeInput

respond_to_auth_challenge_input

See RespondToAuthChallengeInput

revoke_token_input

See RevokeTokenInput

set_risk_configuration_input

See SetRiskConfigurationInput

set_ui_customization_input

See SetUiCustomizationInput

set_user_mfa_preference_input

See SetUserMfaPreferenceInput

set_user_pool_mfa_config_input

See SetUserPoolMfaConfigInput

set_user_settings_input

See SetUserSettingsInput

sign_up_input

See SignUpInput

start_user_import_job_input

See StartUserImportJobInput

stop_user_import_job_input

See StopUserImportJobInput

tag_resource_input

See TagResourceInput

untag_resource_input

See UntagResourceInput

update_auth_event_feedback_input

See UpdateAuthEventFeedbackInput

update_device_status_input

See UpdateDeviceStatusInput

update_group_input

See UpdateGroupInput

update_identity_provider_input

See UpdateIdentityProviderInput

update_resource_server_input

See UpdateResourceServerInput

update_user_attributes_input

See UpdateUserAttributesInput

update_user_pool_client_input

See UpdateUserPoolClientInput

update_user_pool_domain_input

See UpdateUserPoolDomainInput

update_user_pool_input

See UpdateUserPoolInput

verify_software_token_input

See VerifySoftwareTokenInput

verify_user_attribute_input

See VerifyUserAttributeInput

Structs

AddCustomAttributesInput

Represents the request to add custom attributes.

AdminAddUserToGroupInput

AdminConfirmSignUpInput

Represents the request to confirm user registration.

AdminCreateUserInput

Represents the request to create a user in the specified user pool.

AdminDeleteUserAttributesInput

Represents the request to delete user attributes as an administrator.

AdminDeleteUserInput

Represents the request to delete a user as an administrator.

AdminDisableProviderForUserInput

AdminDisableUserInput

Represents the request to disable the user as an administrator.

AdminEnableUserInput

Represents the request that enables the user as an administrator.

AdminForgetDeviceInput

Sends the forgot device request, as an administrator.

AdminGetDeviceInput

Represents the request to get the device, as an administrator.

AdminGetUserInput

Represents the request to get the specified user as an administrator.

AdminInitiateAuthInput

Initiates the authorization request, as an administrator.

AdminLinkProviderForUserInput

AdminListDevicesInput

Represents the request to list devices, as an administrator.

AdminListGroupsForUserInput

AdminListUserAuthEventsInput

AdminRemoveUserFromGroupInput

AdminResetUserPasswordInput

Represents the request to reset a user's password as an administrator.

AdminRespondToAuthChallengeInput

The request to respond to the authentication challenge, as an administrator.

AdminSetUserMfaPreferenceInput

AdminSetUserPasswordInput

AdminSetUserSettingsInput

You can use this parameter to set an MFA configuration that uses the SMS delivery

AdminUpdateAuthEventFeedbackInput

AdminUpdateDeviceStatusInput

The request to update the device status, as an administrator.

AdminUpdateUserAttributesInput

Represents the request to update the user's attributes as an administrator.

AdminUserGlobalSignOutInput

The request to sign out of all devices, as an administrator.

AssociateSoftwareTokenInput

ChangePasswordInput

Represents the request to change a user password.

ConfirmDeviceInput

Confirms the device request.

ConfirmForgotPasswordInput

The request representing the confirmation for a password reset.

ConfirmSignUpInput

Represents the request to confirm registration of a user.

CreateGroupInput

CreateIdentityProviderInput

CreateResourceServerInput

CreateUserImportJobInput

Represents the request to create the user import job.

CreateUserPoolClientInput

Represents the request to create a user pool client.

CreateUserPoolDomainInput

CreateUserPoolInput

Represents the request to create a user pool.

DeleteGroupInput

DeleteIdentityProviderInput

DeleteResourceServerInput

DeleteUserAttributesInput

Represents the request to delete user attributes.

DeleteUserInput

Represents the request to delete a user.

DeleteUserPoolClientInput

Represents the request to delete a user pool client.

DeleteUserPoolDomainInput

DeleteUserPoolInput

Represents the request to delete a user pool.

DescribeIdentityProviderInput

DescribeResourceServerInput

DescribeRiskConfigurationInput

DescribeUserImportJobInput

Represents the request to describe the user import job.

DescribeUserPoolClientInput

Represents the request to describe a user pool client.

DescribeUserPoolDomainInput

DescribeUserPoolInput

Represents the request to describe the user pool.

ForgetDeviceInput

Represents the request to forget the device.

ForgotPasswordInput

Represents the request to reset a user's password.

GetCsvHeaderInput

Represents the request to get the header information for the .csv file for the user

GetDeviceInput

Represents the request to get the device.

GetGroupInput

GetIdentityProviderByIdentifierInput

GetSigningCertificateInput

Request to get a signing certificate from Cognito.

GetUiCustomizationInput

GetUserAttributeVerificationCodeInput

Represents the request to get user attribute verification.

GetUserInput

Represents the request to get information about the user.

GetUserPoolMfaConfigInput

GlobalSignOutInput

Represents the request to sign out all devices.

InitiateAuthInput

Initiates the authentication request.

ListDevicesInput

Represents the request to list the devices.

ListGroupsInput

ListIdentityProvidersInput

ListResourceServersInput

ListTagsForResourceInput

ListUserImportJobsInput

Represents the request to list the user import jobs.

ListUserPoolClientsInput

Represents the request to list the user pool clients.

ListUserPoolsInput

Represents the request to list user pools.

ListUsersInGroupInput

ListUsersInput

Represents the request to list users.

ResendConfirmationCodeInput

Represents the request to resend the confirmation code.

RespondToAuthChallengeInput

The request to respond to an authentication challenge.

RevokeTokenInput

SetRiskConfigurationInput

SetUiCustomizationInput

SetUserMfaPreferenceInput

SetUserPoolMfaConfigInput

SetUserSettingsInput

Represents the request to set user settings.

SignUpInput

Represents the request to register a user.

StartUserImportJobInput

Represents the request to start the user import job.

StopUserImportJobInput

Represents the request to stop the user import job.

TagResourceInput

UntagResourceInput

UpdateAuthEventFeedbackInput

UpdateDeviceStatusInput

Represents the request to update the device status.

UpdateGroupInput

UpdateIdentityProviderInput

UpdateResourceServerInput

UpdateUserAttributesInput

Represents the request to update user attributes.

UpdateUserPoolClientInput

Represents the request to update the user pool client.

UpdateUserPoolDomainInput

The UpdateUserPoolDomain request input.

UpdateUserPoolInput

Represents the request to update the user pool.

VerifySoftwareTokenInput

VerifyUserAttributeInput

Represents the request to verify user attributes.