Struct aws_sdk_cognitoidentityprovider::input::CreateUserPoolClientInput

#[non_exhaustive]
pub struct CreateUserPoolClientInput {

    pub user_pool_id: Option<String>,
    pub client_name: Option<String>,
    pub generate_secret: bool,
    pub refresh_token_validity: i32,
    pub access_token_validity: Option<i32>,
    pub id_token_validity: Option<i32>,
    pub token_validity_units: Option<TokenValidityUnitsType>,
    pub read_attributes: Option<Vec<String>>,
    pub write_attributes: Option<Vec<String>>,
    pub explicit_auth_flows: Option<Vec<ExplicitAuthFlowsType>>,
    pub supported_identity_providers: Option<Vec<String>>,
    pub callback_ur_ls: Option<Vec<String>>,
    pub logout_ur_ls: Option<Vec<String>>,
    pub default_redirect_uri: Option<String>,
    pub allowed_o_auth_flows: Option<Vec<OAuthFlowType>>,
    pub allowed_o_auth_scopes: Option<Vec<String>>,
    pub allowed_o_auth_flows_user_pool_client: bool,
    pub analytics_configuration: Option<AnalyticsConfigurationType>,
    pub prevent_user_existence_errors: Option<PreventUserExistenceErrorTypes>,
    pub enable_token_revocation: Option<bool>,

}

Represents the request to create a user pool client.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

user_pool_id: Option<String>

The user pool ID for the user pool where you want to create a user pool client.

client_name: Option<String>

The client name for the user pool client you would like to create.

generate_secret: bool

Boolean to specify whether you want to generate a secret for the user pool client being created.

refresh_token_validity: i32

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

access_token_validity: Option<i32>

The time limit, between 5 minutes and 1 day, after which the access token is no longer valid and cannot be used. This value will be overridden if you have entered a value in TokenValidityUnits.

id_token_validity: Option<i32>

The time limit, between 5 minutes and 1 day, after which the ID token is no longer valid and cannot be used. This value will be overridden if you have entered a value in TokenValidityUnits.

token_validity_units: Option<TokenValidityUnitsType>

The units in which the validity times are represented in. Default for RefreshToken is days, and default for ID and access tokens are hours.

read_attributes: Option<Vec<String>>

The read attributes.

write_attributes: Option<Vec<String>>

The user pool attributes that the app client can write to.

If your app client allows users to sign in through an identity provider, this array must include all attributes that are mapped to identity provider attributes. Amazon Cognito updates mapped attributes when users sign in to your application through an identity provider. If your app client lacks write access to a mapped attribute, Amazon Cognito throws an error when it attempts to update the attribute. For more information, see Specifying Identity Provider Attribute Mappings for Your User Pool.

explicit_auth_flows: Option<Vec<ExplicitAuthFlowsType>>

The authentication flows that are supported by the user pool clients. Flow names without the ALLOW_ prefix are deprecated in favor of new names with the ALLOW_ prefix. Note that values with ALLOW_ prefix cannot be used along with values without ALLOW_ prefix.

Valid values include:

ALLOW_ADMIN_USER_PASSWORD_AUTH: Enable admin based user password authentication flow ADMIN_USER_PASSWORD_AUTH. This setting replaces the ADMIN_NO_SRP_AUTH setting. With this authentication flow, Cognito receives the password in the request instead of using the SRP (Secure Remote Password protocol) protocol to verify passwords.

ALLOW_CUSTOM_AUTH: Enable Lambda trigger based authentication.

ALLOW_USER_PASSWORD_AUTH: Enable user password-based authentication. In this flow, Cognito receives the password in the request instead of using the SRP protocol to verify passwords.

ALLOW_USER_SRP_AUTH: Enable SRP based authentication.

ALLOW_REFRESH_TOKEN_AUTH: Enable authflow to refresh tokens.

supported_identity_providers: Option<Vec<String>>

A list of provider names for the identity providers that are supported on this client. The following are supported: COGNITO, Facebook, Google and LoginWithAmazon.

callback_ur_ls: Option<Vec<String>>

A list of allowed redirect (callback) URLs for the identity providers.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

logout_ur_ls: Option<Vec<String>>

A list of allowed logout URLs for the identity providers.

default_redirect_uri: Option<String>

The default redirect URI. Must be in the CallbackURLs list.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

allowed_o_auth_flows: Option<Vec<OAuthFlowType>>

The allowed OAuth flows.

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint.

Set to implicit to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

Set to client_credentials to specify that the client should get the access token (and, optionally, ID token, based on scopes) from the token endpoint using a combination of client and client_secret.

allowed_o_auth_scopes: Option<Vec<String>>

The allowed OAuth scopes. Possible values provided by OAuth are: phone, email, openid, and profile. Possible values provided by AWS are: aws.cognito.signin.user.admin. Custom scopes created in Resource Servers are also supported.

allowed_o_auth_flows_user_pool_client: bool

Set to true if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

analytics_configuration: Option<AnalyticsConfigurationType>

The Amazon Pinpoint analytics configuration for collecting metrics for this user pool.

In regions where Pinpoint is not available, Cognito User Pools only supports sending events to Amazon Pinpoint projects in us-east-1. In regions where Pinpoint is available, Cognito User Pools will support sending events to Amazon Pinpoint projects within that same region.

prevent_user_existence_errors: Option<PreventUserExistenceErrorTypes>

Use this setting to choose which errors and responses are returned by Cognito APIs during authentication, account confirmation, and password recovery when the user does not exist in the user pool. When set to ENABLED and the user does not exist, authentication returns an error indicating either the username or password was incorrect, and account confirmation and password recovery return a response indicating a code was sent to a simulated destination. When set to LEGACY, those APIs will return a UserNotFoundException exception if the user does not exist in the user pool.

Valid values include:

ENABLED - This prevents user existence-related errors.

LEGACY - This represents the old behavior of Cognito where user existence related errors are not prevented.

After February 15th 2020, the value of PreventUserExistenceErrors will default to ENABLED for newly created user pool clients if no value is provided.

enable_token_revocation: Option<bool>

Enables or disables token revocation. For more information about revoking tokens, see RevokeToken.

If you don't include this parameter, token revocation is automatically enabled for the new user pool client.

Implementations

impl CreateUserPoolClientInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<CreateUserPoolClient, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateUserPoolClient>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateUserPoolClientInput