Struct aws_sdk_cognitoidentityprovider::input::update_user_pool_client_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UpdateUserPoolClientInput

Implementations

impl Builder

pub fn user_pool_id(self, input: impl Into<String>) -> Self

The user pool ID for the user pool where you want to update the user pool client.

pub fn set_user_pool_id(self, input: Option<String>) -> Self

pub fn client_id(self, input: impl Into<String>) -> Self

The ID of the client associated with the user pool.

pub fn set_client_id(self, input: Option<String>) -> Self

pub fn client_name(self, input: impl Into<String>) -> Self

The client name from the update user pool client request.

pub fn set_client_name(self, input: Option<String>) -> Self

pub fn refresh_token_validity(self, input: i32) -> Self

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

pub fn set_refresh_token_validity(self, input: Option<i32>) -> Self

pub fn access_token_validity(self, input: i32) -> Self

The time limit, after which the access token is no longer valid and cannot be used.

pub fn set_access_token_validity(self, input: Option<i32>) -> Self

pub fn id_token_validity(self, input: i32) -> Self

The time limit, after which the ID token is no longer valid and cannot be used.

pub fn set_id_token_validity(self, input: Option<i32>) -> Self

pub fn token_validity_units(self, input: TokenValidityUnitsType) -> Self

The units in which the validity times are represented in. Default for RefreshToken is days, and default for ID and access tokens are hours.

pub fn set_token_validity_units(
self,
input: Option<TokenValidityUnitsType>
) -> Self

pub fn read_attributes(self, input: impl Into<String>) -> Self

pub fn set_read_attributes(self, input: Option<Vec<String>>) -> Self

pub fn write_attributes(self, input: impl Into<String>) -> Self

pub fn set_write_attributes(self, input: Option<Vec<String>>) -> Self

pub fn explicit_auth_flows(
self,
input: impl Into<ExplicitAuthFlowsType>
) -> Self

pub fn set_explicit_auth_flows(
self,
input: Option<Vec<ExplicitAuthFlowsType>>
) -> Self

pub fn supported_identity_providers(self, input: impl Into<String>) -> Self

pub fn set_supported_identity_providers(
self,
input: Option<Vec<String>>
) -> Self

pub fn callback_ur_ls(self, input: impl Into<String>) -> Self

pub fn set_callback_ur_ls(self, input: Option<Vec<String>>) -> Self

pub fn logout_ur_ls(self, input: impl Into<String>) -> Self

pub fn set_logout_ur_ls(self, input: Option<Vec<String>>) -> Self

pub fn default_redirect_uri(self, input: impl Into<String>) -> Self

The default redirect URI. Must be in the CallbackURLs list.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

pub fn set_default_redirect_uri(self, input: Option<String>) -> Self

pub fn allowed_o_auth_flows(self, input: impl Into<OAuthFlowType>) -> Self

pub fn set_allowed_o_auth_flows(self, input: Option<Vec<OAuthFlowType>>) -> Self

pub fn allowed_o_auth_scopes(self, input: impl Into<String>) -> Self

pub fn set_allowed_o_auth_scopes(self, input: Option<Vec<String>>) -> Self

pub fn allowed_o_auth_flows_user_pool_client(self, input: bool) -> Self

Set to true if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

pub fn set_allowed_o_auth_flows_user_pool_client(
self,
input: Option<bool>
) -> Self

pub fn analytics_configuration(self, input: AnalyticsConfigurationType) -> Self

The Amazon Pinpoint analytics configuration for collecting metrics for this user pool.

In regions where Pinpoint is not available, Cognito User Pools only supports sending events to Amazon Pinpoint projects in us-east-1. In regions where Pinpoint is available, Cognito User Pools will support sending events to Amazon Pinpoint projects within that same region.

pub fn set_analytics_configuration(
self,
input: Option<AnalyticsConfigurationType>
) -> Self

pub fn prevent_user_existence_errors(
self,
input: PreventUserExistenceErrorTypes
) -> Self

Use this setting to choose which errors and responses are returned by Cognito APIs during authentication, account confirmation, and password recovery when the user does not exist in the user pool. When set to ENABLED and the user does not exist, authentication returns an error indicating either the username or password was incorrect, and account confirmation and password recovery return a response indicating a code was sent to a simulated destination. When set to LEGACY, those APIs will return a UserNotFoundException exception if the user does not exist in the user pool.

Valid values include:

ENABLED - This prevents user existence-related errors.

LEGACY - This represents the old behavior of Cognito where user existence related errors are not prevented.

After February 15th 2020, the value of PreventUserExistenceErrors will default to ENABLED for newly created user pool clients if no value is provided.

pub fn set_prevent_user_existence_errors(
self,
input: Option<PreventUserExistenceErrorTypes>
) -> Self

pub fn enable_token_revocation(self, input: bool) -> Self

Enables or disables token revocation. For more information about revoking tokens, see RevokeToken.

pub fn set_enable_token_revocation(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<UpdateUserPoolClientInput, BuildError>

Consumes the builder and constructs a UpdateUserPoolClientInput