Module aws_sdk_cognitoidentityprovider::model

Modules

account_recovery_setting_type

See AccountRecoverySettingType

account_takeover_action_type

See AccountTakeoverActionType

account_takeover_actions_type

See AccountTakeoverActionsType

account_takeover_risk_configuration_type

See AccountTakeoverRiskConfigurationType

admin_create_user_config_type

See AdminCreateUserConfigType

analytics_configuration_type

See AnalyticsConfigurationType

analytics_metadata_type

See AnalyticsMetadataType

attribute_type

See AttributeType

auth_event_type

See AuthEventType

authentication_result_type

See AuthenticationResultType

challenge_response_type

See ChallengeResponseType

code_delivery_details_type

See CodeDeliveryDetailsType

compromised_credentials_actions_type

See CompromisedCredentialsActionsType

compromised_credentials_risk_configuration_type

See CompromisedCredentialsRiskConfigurationType

context_data_type

See ContextDataType

custom_domain_config_type

See CustomDomainConfigType

custom_email_lambda_version_config_type

See CustomEmailLambdaVersionConfigType

custom_sms_lambda_version_config_type

See CustomSmsLambdaVersionConfigType

device_configuration_type

See DeviceConfigurationType

device_secret_verifier_config_type

See DeviceSecretVerifierConfigType

device_type

See DeviceType

domain_description_type

See DomainDescriptionType

email_configuration_type

See EmailConfigurationType

event_context_data_type

See EventContextDataType

event_feedback_type

See EventFeedbackType

event_risk_type

See EventRiskType

group_type

See GroupType

http_header

See HttpHeader

identity_provider_type

See IdentityProviderType

lambda_config_type

See LambdaConfigType

message_template_type

See MessageTemplateType

mfa_option_type

See MfaOptionType

new_device_metadata_type

See NewDeviceMetadataType

notify_configuration_type

See NotifyConfigurationType

notify_email_type

See NotifyEmailType

number_attribute_constraints_type

See NumberAttributeConstraintsType

password_policy_type

See PasswordPolicyType

provider_description

See ProviderDescription

provider_user_identifier_type

See ProviderUserIdentifierType

recovery_option_type

See RecoveryOptionType

resource_server_scope_type

See ResourceServerScopeType

resource_server_type

See ResourceServerType

risk_configuration_type

See RiskConfigurationType

risk_exception_configuration_type

See RiskExceptionConfigurationType

schema_attribute_type

See SchemaAttributeType

sms_configuration_type

See SmsConfigurationType

sms_mfa_config_type

See SmsMfaConfigType

sms_mfa_settings_type

See SmsMfaSettingsType

software_token_mfa_config_type

See SoftwareTokenMfaConfigType

software_token_mfa_settings_type

See SoftwareTokenMfaSettingsType

string_attribute_constraints_type

See StringAttributeConstraintsType

token_validity_units_type

See TokenValidityUnitsType

ui_customization_type

See UiCustomizationType

user_context_data_type

See UserContextDataType

user_import_job_type

See UserImportJobType

user_pool_add_ons_type

See UserPoolAddOnsType

user_pool_client_description

See UserPoolClientDescription

user_pool_client_type

See UserPoolClientType

user_pool_description_type

See UserPoolDescriptionType

user_pool_policy_type

See UserPoolPolicyType

user_pool_type

See UserPoolType

user_type

See UserType

username_configuration_type

See UsernameConfigurationType

verification_message_template_type

See VerificationMessageTemplateType

Structs

AccountRecoverySettingType

The data type for AccountRecoverySetting.

AccountTakeoverActionType

Account takeover action type.

AccountTakeoverActionsType

Account takeover actions type.

AccountTakeoverRiskConfigurationType

Configuration for mitigation actions and notification for different levels of risk

AdminCreateUserConfigType

The configuration for creating a new user profile.

AnalyticsConfigurationType

The Amazon Pinpoint analytics configuration for collecting metrics for a user

AnalyticsMetadataType

An Amazon Pinpoint analytics endpoint.

AttributeType

Specifies whether the attribute is standard or custom.

AuthEventType

The authentication event type.

AuthenticationResultType

The authentication result.

ChallengeResponseType

The challenge response type.

CodeDeliveryDetailsType

The code delivery details being returned from the server.

CompromisedCredentialsActionsType

The compromised credentials actions type

CompromisedCredentialsRiskConfigurationType

The compromised credentials risk configuration type.

ContextDataType

Contextual user data type used for evaluating the risk of an unexpected event by

CustomDomainConfigType

The configuration for a custom domain that hosts the sign-up and sign-in webpages for

CustomEmailLambdaVersionConfigType

A custom email sender Lambda configuration type.

CustomSmsLambdaVersionConfigType

A custom SMS sender Lambda configuration type.

DeviceConfigurationType

The configuration for the user pool's device tracking.

DeviceSecretVerifierConfigType

The device verifier against which it will be authenticated.

DeviceType

The device type.

DomainDescriptionType

A container for information about a domain.

EmailConfigurationType

The email configuration type.

EventContextDataType

Specifies the user context data captured at the time of an event request.

EventFeedbackType

Specifies the event feedback type.

EventRiskType

The event risk type.

GroupType

The group type.

HttpHeader

The HTTP header.

IdentityProviderType

A container for information about an identity provider.

LambdaConfigType

Specifies the configuration for AWS Lambda triggers.

MessageTemplateType

The message template structure.

MfaOptionType

NewDeviceMetadataType

The new device metadata type.

NotifyConfigurationType

The notify configuration type.

NotifyEmailType

The notify email type.

NumberAttributeConstraintsType

The minimum and maximum value of an attribute that is of the number data type.

PasswordPolicyType

The password policy type.

ProviderDescription

A container for identity provider details.

ProviderUserIdentifierType

A container for information about an identity provider for a user pool.

RecoveryOptionType

A map containing a priority as a key, and recovery method name as a value.

ResourceServerScopeType

A resource server scope.

ResourceServerType

A container for information about a resource server for a user pool.

RiskConfigurationType

The risk configuration type.

RiskExceptionConfigurationType

The type of the configuration to override the risk decision.

SchemaAttributeType

Contains information about the schema attribute.

SmsConfigurationType

The SMS configuration type that includes the settings the Cognito User Pool needs to

SmsMfaConfigType

The SMS text message multi-factor authentication (MFA) configuration type.

SmsMfaSettingsType

The type used for enabling SMS MFA at the user level. Phone numbers don't need to be

SoftwareTokenMfaConfigType

The type used for enabling software token MFA at the user pool level.

SoftwareTokenMfaSettingsType

The type used for enabling software token MFA at the user level. If an MFA type is

StringAttributeConstraintsType

The constraints associated with a string attribute.

TokenValidityUnitsType

The data type for TokenValidityUnits that specifics the time measurements for token

UiCustomizationType

A container for the UI customization information for a user pool's built-in app

UserContextDataType

Contextual data such as the user's device fingerprint, IP address, or location used

UserImportJobType

The user import job type.

UserPoolAddOnsType

The user pool add-ons type.

UserPoolClientDescription

The description of the user pool client.

UserPoolClientType

Contains information about a user pool client.

UserPoolDescriptionType

A user pool description.

UserPoolPolicyType

The policy associated with a user pool.

UserPoolType

A container for information about the user pool.

UserType

The user type.

UsernameConfigurationType

The username configuration type.

VerificationMessageTemplateType

The template for verification messages.

Enums

AccountTakeoverEventActionType

AdvancedSecurityModeType

AliasAttributeType

AttributeDataType

AuthFlowType

ChallengeName

ChallengeNameType

ChallengeResponse

CompromisedCredentialsEventActionType

CustomEmailSenderLambdaVersionType

CustomSmsSenderLambdaVersionType

DefaultEmailOptionType

DeliveryMediumType

DeviceRememberedStatusType

DomainStatusType

EmailSendingAccountType

EventFilterType

EventResponseType

EventType

ExplicitAuthFlowsType

FeedbackValueType

IdentityProviderTypeType

MessageActionType

OAuthFlowType

PreventUserExistenceErrorTypes

RecoveryOptionNameType

RiskDecisionType

RiskLevelType

StatusType

TimeUnitsType

UserImportJobStatusType

UserPoolMfaType

UserStatusType

NOTE: UserStatusType::Unknown has been renamed to ::UnknownValue.

UsernameAttributeType

VerifiedAttributeType

VerifySoftwareTokenResponseType