Struct aws_sdk_cognitoidentityprovider::model::user_pool_type::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UserPoolType

Implementations

impl Builder

pub fn id(self, input: impl Into<String>) -> Self

The ID of the user pool.

pub fn set_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name of the user pool.

pub fn set_name(self, input: Option<String>) -> Self

pub fn policies(self, input: UserPoolPolicyType) -> Self

The policies associated with the user pool.

pub fn set_policies(self, input: Option<UserPoolPolicyType>) -> Self

pub fn lambda_config(self, input: LambdaConfigType) -> Self

The AWS Lambda triggers associated with the user pool.

pub fn set_lambda_config(self, input: Option<LambdaConfigType>) -> Self

pub fn status(self, input: StatusType) -> Self

The status of a user pool.

pub fn set_status(self, input: Option<StatusType>) -> Self

pub fn last_modified_date(self, input: Instant) -> Self

The date the user pool was last modified.

pub fn set_last_modified_date(self, input: Option<Instant>) -> Self

pub fn creation_date(self, input: Instant) -> Self

The date the user pool was created.

pub fn set_creation_date(self, input: Option<Instant>) -> Self

pub fn schema_attributes(self, input: impl Into<SchemaAttributeType>) -> Self

pub fn set_schema_attributes(
self,
input: Option<Vec<SchemaAttributeType>>
) -> Self

pub fn auto_verified_attributes(
self,
input: impl Into<VerifiedAttributeType>
) -> Self

pub fn set_auto_verified_attributes(
self,
input: Option<Vec<VerifiedAttributeType>>
) -> Self

pub fn alias_attributes(self, input: impl Into<AliasAttributeType>) -> Self

pub fn set_alias_attributes(
self,
input: Option<Vec<AliasAttributeType>>
) -> Self

pub fn username_attributes(
self,
input: impl Into<UsernameAttributeType>
) -> Self

pub fn set_username_attributes(
self,
input: Option<Vec<UsernameAttributeType>>
) -> Self

pub fn sms_verification_message(self, input: impl Into<String>) -> Self

The contents of the SMS verification message.

pub fn set_sms_verification_message(self, input: Option<String>) -> Self

pub fn email_verification_message(self, input: impl Into<String>) -> Self

The contents of the email verification message.

pub fn set_email_verification_message(self, input: Option<String>) -> Self

pub fn email_verification_subject(self, input: impl Into<String>) -> Self

The subject of the email verification message.

pub fn set_email_verification_subject(self, input: Option<String>) -> Self

pub fn verification_message_template(
self,
input: VerificationMessageTemplateType
) -> Self

The template for verification messages.

pub fn set_verification_message_template(
self,
input: Option<VerificationMessageTemplateType>
) -> Self

pub fn sms_authentication_message(self, input: impl Into<String>) -> Self

The contents of the SMS authentication message.

pub fn set_sms_authentication_message(self, input: Option<String>) -> Self

pub fn mfa_configuration(self, input: UserPoolMfaType) -> Self

Can be one of the following values:

OFF - MFA tokens are not required and cannot be specified during user registration.

ON - MFA tokens are required for all user registrations. You can only specify required when you are initially creating a user pool.

OPTIONAL - Users have the option when registering to create an MFA token.

pub fn set_mfa_configuration(self, input: Option<UserPoolMfaType>) -> Self

pub fn device_configuration(self, input: DeviceConfigurationType) -> Self

The device configuration.

pub fn set_device_configuration(
self,
input: Option<DeviceConfigurationType>
) -> Self

pub fn estimated_number_of_users(self, input: i32) -> Self

A number estimating the size of the user pool.

pub fn set_estimated_number_of_users(self, input: Option<i32>) -> Self

pub fn email_configuration(self, input: EmailConfigurationType) -> Self

The email configuration.

pub fn set_email_configuration(
self,
input: Option<EmailConfigurationType>
) -> Self

pub fn sms_configuration(self, input: SmsConfigurationType) -> Self

The SMS configuration.

pub fn set_sms_configuration(self, input: Option<SmsConfigurationType>) -> Self

pub fn user_pool_tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_user_pool_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn sms_configuration_failure(self, input: impl Into<String>) -> Self

The reason why the SMS configuration cannot send the messages to your users.

This message might include comma-separated values to describe why your SMS configuration can't send messages to user pool end users.

InvalidSmsRoleAccessPolicyException - The IAM role which Cognito uses to send SMS messages is not properly configured. For more information, see SmsConfigurationType.

SNSSandbox - The AWS account is in SNS Sandbox and messages won’t reach unverified end users. This parameter won’t get populated with SNSSandbox if the IAM user creating the user pool doesn’t have SNS permissions. To learn how to move your AWS account out of the sandbox, see Moving out of the SMS sandbox.

pub fn set_sms_configuration_failure(self, input: Option<String>) -> Self

pub fn email_configuration_failure(self, input: impl Into<String>) -> Self

The reason why the email configuration cannot send the messages to your users.

pub fn set_email_configuration_failure(self, input: Option<String>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

Holds the domain prefix if the user pool has a domain associated with it.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn custom_domain(self, input: impl Into<String>) -> Self

A custom domain name that you provide to Amazon Cognito. This parameter applies only if you use a custom domain to host the sign-up and sign-in pages for your application. For example: auth.example.com.

For more information about adding a custom domain to your user pool, see Using Your Own Domain for the Hosted UI.

pub fn set_custom_domain(self, input: Option<String>) -> Self

pub fn admin_create_user_config(self, input: AdminCreateUserConfigType) -> Self

The configuration for AdminCreateUser requests.

pub fn set_admin_create_user_config(
self,
input: Option<AdminCreateUserConfigType>
) -> Self

pub fn user_pool_add_ons(self, input: UserPoolAddOnsType) -> Self

The user pool add-ons.

pub fn set_user_pool_add_ons(self, input: Option<UserPoolAddOnsType>) -> Self

pub fn username_configuration(self, input: UsernameConfigurationType) -> Self

You can choose to enable case sensitivity on the username input for the selected sign-in option. For example, when this is set to False, users will be able to sign in using either "username" or "Username". This configuration is immutable once it has been set. For more information, see UsernameConfigurationType.

pub fn set_username_configuration(
self,
input: Option<UsernameConfigurationType>
) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the user pool.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn account_recovery_setting(self, input: AccountRecoverySettingType) -> Self

Use this setting to define which verified available method a user can use to recover their password when they call ForgotPassword. It allows you to define a preferred method when a user has more than one method available. With this setting, SMS does not qualify for a valid password recovery mechanism if the user also has SMS MFA enabled. In the absence of this setting, Cognito uses the legacy behavior to determine the recovery method where SMS is preferred over email.

pub fn set_account_recovery_setting(
self,
input: Option<AccountRecoverySettingType>
) -> Self

pub fn build(self) -> UserPoolType

Consumes the builder and constructs a UserPoolType