Module aws_sdk_cognitoidentityprovider::operation

Structs

AddCustomAttributes

Adds additional user attributes to the user pool schema.

AdminAddUserToGroup

Adds the specified user to the specified group.

AdminConfirmSignUp

Confirms user registration as an admin without using a confirmation code. Works on any

AdminCreateUser

Creates a new user in the specified user pool.

AdminDeleteUser

Deletes a user as an administrator. Works on any user.

AdminDeleteUserAttributes

Deletes the user attributes in a user pool as an administrator. Works on any

AdminDisableProviderForUser

Disables the user from signing in with the specified external (SAML or social)

AdminDisableUser

Disables the specified user.

AdminEnableUser

Enables the specified user as an administrator. Works on any user.

AdminForgetDevice

Forgets the device, as an administrator.

AdminGetDevice

Gets the device, as an administrator.

AdminGetUser

Gets the specified user by user name in a user pool as an administrator. Works on any

AdminInitiateAuth

Initiates the authentication flow, as an administrator.

AdminLinkProviderForUser

Links an existing user account in a user pool (DestinationUser) to an

AdminListDevices

Lists devices, as an administrator.

AdminListGroupsForUser

Lists the groups that the user belongs to.

AdminListUserAuthEvents

Lists a history of user activity and any risks detected as part of Amazon Cognito

AdminRemoveUserFromGroup

Removes the specified user from the specified group.

AdminResetUserPassword

Resets the specified user's password in a user pool as an administrator. Works on any

AdminRespondToAuthChallenge

Responds to an authentication challenge, as an administrator.

AdminSetUserMFAPreference

Sets the user's multi-factor authentication (MFA) preference, including which MFA

AdminSetUserPassword

Sets the specified user's password in a user pool as an administrator. Works on any

AdminSetUserSettings

AdminUpdateAuthEventFeedback

Provides feedback for an authentication event as to whether it was from a valid user.

AdminUpdateDeviceStatus

Updates the device status as an administrator.

AdminUpdateUserAttributes

Updates the specified user's attributes, including developer attributes, as an

AdminUserGlobalSignOut

Signs out users from all devices, as an administrator. It also invalidates all refresh

AssociateSoftwareToken

Returns a unique generated shared secret key code for the user account. The request

ChangePassword

Changes the password for a specified user in a user pool.

ConfirmDevice

Confirms tracking of the device. This API call is the call that begins device

ConfirmForgotPassword

Allows a user to enter a confirmation code to reset a forgotten password.

ConfirmSignUp

Confirms registration of a user and handles the existing alias from a previous

CreateGroup

Creates a new group in the specified user pool.

CreateIdentityProvider

Creates an identity provider for a user pool.

CreateResourceServer

Creates a new OAuth2.0 resource server and defines custom scopes in it.

CreateUserImportJob

Creates the user import job.

CreateUserPool

Creates a new Amazon Cognito user pool and sets the password policy for the

CreateUserPoolClient

Creates the user pool client.

CreateUserPoolDomain

Creates a new domain for a user pool.

DeleteGroup

Deletes a group.

DeleteIdentityProvider

Deletes an identity provider for a user pool.

DeleteResourceServer

Deletes a resource server.

DeleteUser

Allows a user to delete himself or herself.

DeleteUserAttributes

Deletes the attributes for a user.

DeleteUserPool

Deletes the specified Amazon Cognito user pool.

DeleteUserPoolClient

Allows the developer to delete the user pool client.

DeleteUserPoolDomain

Deletes a domain for a user pool.

DescribeIdentityProvider

Gets information about a specific identity provider.

DescribeResourceServer

Describes a resource server.

DescribeRiskConfiguration

Describes the risk configuration.

DescribeUserImportJob

Describes the user import job.

DescribeUserPool

Returns the configuration information and metadata of the specified user pool.

DescribeUserPoolClient

Client method for returning the configuration information and metadata of the

DescribeUserPoolDomain

Gets information about a domain.

ForgetDevice

Forgets the specified device.

ForgotPassword

Calling this API causes a message to be sent to the end user with a confirmation code

GetCSVHeader

Gets the header information for the .csv file to be used as input for the user import

GetDevice

Gets the device.

GetGroup

Gets a group.

GetIdentityProviderByIdentifier

Gets the specified identity provider.

GetSigningCertificate

This method takes a user pool ID, and returns the signing certificate.

GetUICustomization

Gets the UI Customization information for a particular app client's app UI, if there

GetUser

Gets the user attributes and metadata for a user.

GetUserAttributeVerificationCode

Gets the user attribute verification code for the specified attribute name.

GetUserPoolMfaConfig

Gets the user pool multi-factor authentication (MFA) configuration.

GlobalSignOut

Signs out users from all devices. It also invalidates all refresh tokens issued to a

InitiateAuth

Initiates the authentication flow.

ListDevices

Lists the devices.

ListGroups

Lists the groups associated with a user pool.

ListIdentityProviders

Lists information about all identity providers for a user pool.

ListResourceServers

Lists the resource servers for a user pool.

ListTagsForResource

Lists the tags that are assigned to an Amazon Cognito user pool.

ListUserImportJobs

Lists the user import jobs.

ListUserPoolClients

Lists the clients that have been created for the specified user pool.

ListUserPools

Lists the user pools associated with an AWS account.

ListUsers

Lists the users in the Amazon Cognito user pool.

ListUsersInGroup

Lists the users in the specified group.

ResendConfirmationCode

Resends the confirmation (for confirmation of registration) to a specific user in the

RespondToAuthChallenge

Responds to the authentication challenge.

RevokeToken

Revokes all of the access tokens generated by the specified refresh token. After the token is revoked, you can not

SetRiskConfiguration

Configures actions on detected risks. To delete the risk configuration for

SetUICustomization

Sets the UI customization information for a user pool's built-in app UI.

SetUserMFAPreference

Set the user's multi-factor authentication (MFA) method preference, including which

SetUserPoolMfaConfig

Set the user pool multi-factor authentication (MFA) configuration.

SetUserSettings

SignUp

Registers the user in the specified user pool and creates a user name, password, and

StartUserImportJob

Starts the user import.

StopUserImportJob

Stops the user import job.

TagResource

Assigns a set of tags to an Amazon Cognito user pool. A tag is a label that you can

UntagResource

Removes the specified tags from an Amazon Cognito user pool. You can use this action

UpdateAuthEventFeedback

Provides the feedback for an authentication event whether it was from a valid user or

UpdateDeviceStatus

Updates the device status.

UpdateGroup

Updates the specified group with the specified attributes.

UpdateIdentityProvider

Updates identity provider information for a user pool.

UpdateResourceServer

Updates the name and scopes of resource server. All other fields are read-only.

UpdateUserAttributes

Allows a user to update a specific attribute (one at a time).

UpdateUserPool

Updates the specified user pool with the specified attributes. You can get a list of

UpdateUserPoolClient

Updates the specified user pool app client with the specified attributes. You can get

UpdateUserPoolDomain

Updates the Secure Sockets Layer (SSL) certificate for the custom domain for your user

VerifySoftwareToken

Use this API to register a user's entered TOTP code and mark the user's software token

VerifyUserAttribute

Verifies the specified user attributes in the user pool.