Module aws_sdk_cognitoidentityprovider::output

Modules

add_custom_attributes_output

See AddCustomAttributesOutput

admin_add_user_to_group_output

See AdminAddUserToGroupOutput

admin_confirm_sign_up_output

See AdminConfirmSignUpOutput

admin_create_user_output

See AdminCreateUserOutput

admin_delete_user_attributes_output

See AdminDeleteUserAttributesOutput

admin_delete_user_output

See AdminDeleteUserOutput

admin_disable_provider_for_user_output

See AdminDisableProviderForUserOutput

admin_disable_user_output

See AdminDisableUserOutput

admin_enable_user_output

See AdminEnableUserOutput

admin_forget_device_output

See AdminForgetDeviceOutput

admin_get_device_output

See AdminGetDeviceOutput

admin_get_user_output

See AdminGetUserOutput

admin_initiate_auth_output

See AdminInitiateAuthOutput

admin_link_provider_for_user_output

See AdminLinkProviderForUserOutput

admin_list_devices_output

See AdminListDevicesOutput

admin_list_groups_for_user_output

See AdminListGroupsForUserOutput

admin_list_user_auth_events_output

See AdminListUserAuthEventsOutput

admin_remove_user_from_group_output

See AdminRemoveUserFromGroupOutput

admin_reset_user_password_output

See AdminResetUserPasswordOutput

admin_respond_to_auth_challenge_output

See AdminRespondToAuthChallengeOutput

admin_set_user_mfa_preference_output

See AdminSetUserMfaPreferenceOutput

admin_set_user_password_output

See AdminSetUserPasswordOutput

admin_set_user_settings_output

See AdminSetUserSettingsOutput

admin_update_auth_event_feedback_output

See AdminUpdateAuthEventFeedbackOutput

admin_update_device_status_output

See AdminUpdateDeviceStatusOutput

admin_update_user_attributes_output

See AdminUpdateUserAttributesOutput

admin_user_global_sign_out_output

See AdminUserGlobalSignOutOutput

associate_software_token_output

See AssociateSoftwareTokenOutput

change_password_output

See ChangePasswordOutput

confirm_device_output

See ConfirmDeviceOutput

confirm_forgot_password_output

See ConfirmForgotPasswordOutput

confirm_sign_up_output

See ConfirmSignUpOutput

create_group_output

See CreateGroupOutput

create_identity_provider_output

See CreateIdentityProviderOutput

create_resource_server_output

See CreateResourceServerOutput

create_user_import_job_output

See CreateUserImportJobOutput

create_user_pool_client_output

See CreateUserPoolClientOutput

create_user_pool_domain_output

See CreateUserPoolDomainOutput

create_user_pool_output

See CreateUserPoolOutput

delete_group_output

See DeleteGroupOutput

delete_identity_provider_output

See DeleteIdentityProviderOutput

delete_resource_server_output

See DeleteResourceServerOutput

delete_user_attributes_output

See DeleteUserAttributesOutput

delete_user_output

See DeleteUserOutput

delete_user_pool_client_output

See DeleteUserPoolClientOutput

delete_user_pool_domain_output

See DeleteUserPoolDomainOutput

delete_user_pool_output

See DeleteUserPoolOutput

describe_identity_provider_output

See DescribeIdentityProviderOutput

describe_resource_server_output

See DescribeResourceServerOutput

describe_risk_configuration_output

See DescribeRiskConfigurationOutput

describe_user_import_job_output

See DescribeUserImportJobOutput

describe_user_pool_client_output

See DescribeUserPoolClientOutput

describe_user_pool_domain_output

See DescribeUserPoolDomainOutput

describe_user_pool_output

See DescribeUserPoolOutput

forget_device_output

See ForgetDeviceOutput

forgot_password_output

See ForgotPasswordOutput

get_csv_header_output

See GetCsvHeaderOutput

get_device_output

See GetDeviceOutput

get_group_output

See GetGroupOutput

get_identity_provider_by_identifier_output

See GetIdentityProviderByIdentifierOutput

get_signing_certificate_output

See GetSigningCertificateOutput

get_ui_customization_output

See GetUiCustomizationOutput

get_user_attribute_verification_code_output

See GetUserAttributeVerificationCodeOutput

get_user_output

See GetUserOutput

get_user_pool_mfa_config_output

See GetUserPoolMfaConfigOutput

global_sign_out_output

See GlobalSignOutOutput

initiate_auth_output

See InitiateAuthOutput

list_devices_output

See ListDevicesOutput

list_groups_output

See ListGroupsOutput

list_identity_providers_output

See ListIdentityProvidersOutput

list_resource_servers_output

See ListResourceServersOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

list_user_import_jobs_output

See ListUserImportJobsOutput

list_user_pool_clients_output

See ListUserPoolClientsOutput

list_user_pools_output

See ListUserPoolsOutput

list_users_in_group_output

See ListUsersInGroupOutput

list_users_output

See ListUsersOutput

resend_confirmation_code_output

See ResendConfirmationCodeOutput

respond_to_auth_challenge_output

See RespondToAuthChallengeOutput

revoke_token_output

See RevokeTokenOutput

set_risk_configuration_output

See SetRiskConfigurationOutput

set_ui_customization_output

See SetUiCustomizationOutput

set_user_mfa_preference_output

See SetUserMfaPreferenceOutput

set_user_pool_mfa_config_output

See SetUserPoolMfaConfigOutput

set_user_settings_output

See SetUserSettingsOutput

sign_up_output

See SignUpOutput

start_user_import_job_output

See StartUserImportJobOutput

stop_user_import_job_output

See StopUserImportJobOutput

tag_resource_output

See TagResourceOutput

untag_resource_output

See UntagResourceOutput

update_auth_event_feedback_output

See UpdateAuthEventFeedbackOutput

update_device_status_output

See UpdateDeviceStatusOutput

update_group_output

See UpdateGroupOutput

update_identity_provider_output

See UpdateIdentityProviderOutput

update_resource_server_output

See UpdateResourceServerOutput

update_user_attributes_output

See UpdateUserAttributesOutput

update_user_pool_client_output

See UpdateUserPoolClientOutput

update_user_pool_domain_output

See UpdateUserPoolDomainOutput

update_user_pool_output

See UpdateUserPoolOutput

verify_software_token_output

See VerifySoftwareTokenOutput

verify_user_attribute_output

See VerifyUserAttributeOutput

Structs

AddCustomAttributesOutput

Represents the response from the server for the request to add custom

AdminAddUserToGroupOutput

AdminConfirmSignUpOutput

Represents the response from the server for the request to confirm

AdminCreateUserOutput

Represents the response from the server to the request to create the user.

AdminDeleteUserAttributesOutput

Represents the response received from the server for a request to delete user

AdminDeleteUserOutput

AdminDisableProviderForUserOutput

AdminDisableUserOutput

Represents the response received from the server to disable the user as an

AdminEnableUserOutput

Represents the response from the server for the request to enable a user as an

AdminForgetDeviceOutput

AdminGetDeviceOutput

Gets the device response, as an administrator.

AdminGetUserOutput

Represents the response from the server from the request to get the specified user as

AdminInitiateAuthOutput

Initiates the authentication response, as an administrator.

AdminLinkProviderForUserOutput

AdminListDevicesOutput

Lists the device's response, as an administrator.

AdminListGroupsForUserOutput

AdminListUserAuthEventsOutput

AdminRemoveUserFromGroupOutput

AdminResetUserPasswordOutput

Represents the response from the server to reset a user password as an

AdminRespondToAuthChallengeOutput

Responds to the authentication challenge, as an administrator.

AdminSetUserMfaPreferenceOutput

AdminSetUserPasswordOutput

AdminSetUserSettingsOutput

Represents the response from the server to set user settings as an

AdminUpdateAuthEventFeedbackOutput

AdminUpdateDeviceStatusOutput

The status response from the request to update the device, as an administrator.

AdminUpdateUserAttributesOutput

Represents the response from the server for the request to update user attributes as

AdminUserGlobalSignOutOutput

The global sign-out response, as an administrator.

AssociateSoftwareTokenOutput

ChangePasswordOutput

The response from the server to the change password request.

ConfirmDeviceOutput

Confirms the device response.

ConfirmForgotPasswordOutput

The response from the server that results from a user's request to retrieve a

ConfirmSignUpOutput

Represents the response from the server for the registration confirmation.

CreateGroupOutput

CreateIdentityProviderOutput

CreateResourceServerOutput

CreateUserImportJobOutput

Represents the response from the server to the request to create the user import

CreateUserPoolClientOutput

Represents the response from the server to create a user pool client.

CreateUserPoolDomainOutput

CreateUserPoolOutput

Represents the response from the server for the request to create a user pool.

DeleteGroupOutput

DeleteIdentityProviderOutput

DeleteResourceServerOutput

DeleteUserAttributesOutput

Represents the response from the server to delete user attributes.

DeleteUserOutput

DeleteUserPoolClientOutput

DeleteUserPoolDomainOutput

DeleteUserPoolOutput

DescribeIdentityProviderOutput

DescribeResourceServerOutput

DescribeRiskConfigurationOutput

DescribeUserImportJobOutput

Represents the response from the server to the request to describe the user import

DescribeUserPoolClientOutput

Represents the response from the server from a request to describe the user pool

DescribeUserPoolDomainOutput

DescribeUserPoolOutput

Represents the response to describe the user pool.

ForgetDeviceOutput

ForgotPasswordOutput

Respresents the response from the server regarding the request to reset a

GetCsvHeaderOutput

Represents the response from the server to the request to get the header information

GetDeviceOutput

Gets the device response.

GetGroupOutput

GetIdentityProviderByIdentifierOutput

GetSigningCertificateOutput

Response from Cognito for a signing certificate request.

GetUiCustomizationOutput

GetUserAttributeVerificationCodeOutput

The verification code response returned by the server response to get the user

GetUserOutput

Represents the response from the server from the request to get information about the

GetUserPoolMfaConfigOutput

GlobalSignOutOutput

The response to the request to sign out all devices.

InitiateAuthOutput

Initiates the authentication response.

ListDevicesOutput

Represents the response to list devices.

ListGroupsOutput

ListIdentityProvidersOutput

ListResourceServersOutput

ListTagsForResourceOutput

ListUserImportJobsOutput

Represents the response from the server to the request to list the user import

ListUserPoolClientsOutput

Represents the response from the server that lists user pool clients.

ListUserPoolsOutput

Represents the response to list user pools.

ListUsersInGroupOutput

ListUsersOutput

The response from the request to list users.

ResendConfirmationCodeOutput

The response from the server when the Amazon Cognito Your User Pools service makes the

RespondToAuthChallengeOutput

The response to respond to the authentication challenge.

RevokeTokenOutput

SetRiskConfigurationOutput

SetUiCustomizationOutput

SetUserMfaPreferenceOutput

SetUserPoolMfaConfigOutput

SetUserSettingsOutput

The response from the server for a set user settings request.

SignUpOutput

The response from the server for a registration request.

StartUserImportJobOutput

Represents the response from the server to the request to start the user import

StopUserImportJobOutput

Represents the response from the server to the request to stop the user import

TagResourceOutput

UntagResourceOutput

UpdateAuthEventFeedbackOutput

UpdateDeviceStatusOutput

The response to the request to update the device status.

UpdateGroupOutput

UpdateIdentityProviderOutput

UpdateResourceServerOutput

UpdateUserAttributesOutput

Represents the response from the server for the request to update user

UpdateUserPoolClientOutput

Represents the response from the server to the request to update the user pool

UpdateUserPoolDomainOutput

The UpdateUserPoolDomain response output.

UpdateUserPoolOutput

Represents the response from the server when you make a request to update the user

VerifySoftwareTokenOutput

VerifyUserAttributeOutput

A container representing the response from the server from the request to verify user